requirements.txt
Django<3.1.1
psycopg2-binary>=2.8
djangorestframework>=3.11
django-cors-headers>=3.0
